Update: 障害物をレーン上に生成
## 概要
- 障害物を地面のランダムな位置に生成すると2レーンにまたがる可能性があるため、レーン上に生成されるように変更

## 実装内容
- 障害物生成時、生成位置として取得したランダムな値を整数値に置換

## 動作確認
- 障害物がプレイヤーの移動するレーン上に配置され、レーンをまたいだ配置が起きていないこと

+++ ZigzagCube/Assets/Game/Scripts/ObstacleSpawner.cs
@@ -17,9 +17,9 @@ public class ObstacleSpawner : MonoBehaviour
     {
         // スポーン範囲内からランダムに生成位置を取得
         Vector3 generatePos = new Vector3(
-                Random.Range(spawnArea.x / -2, spawnArea.x / 2),
+                (int)Random.Range(spawnArea.x / -2, spawnArea.x / 2),
                 1f,
-                Random.Range(spawnArea.x / -2, spawnArea.x / 2)
+                (int)Random.Range(spawnArea.x / -2, spawnArea.x / 2)
             );
         // 初回生成時はインスタンス生成と親設定
         if(generatedObstacle == null)
